Remove PowsyblException in LoadTwoTransformersTapChangers
